[Chore] 온보딩클로즈 로직 부모 컴포넌트로 이동
